Fixed issue occurs with 50M inserts because of big file size arithmetic overflow.

+++ src/ZoneTree/Segments/Disk/DecompressedBlock.cs
@@ -73,11 +73,11 @@ public class DecompressedBlock
         return Bytes.AsSpan().Slice(offset, length).ToArray();
     }
 
-    public int TrimRight(int length)
+    public int TrimRight(long length)
     {
         var amount = Math.Min(length, Length);
         var newSize = Length - amount;
-        Bytes = Bytes.AsSpan(0, newSize).ToArray();
-        return amount;
+        Bytes = Bytes.AsSpan(0, (int)newSize).ToArray();
+        return (int)amount;
     }
 }
